need to use double instead of float for data members containing selection parameters

+++ dataSubselector/src/dataSubselector/dataSubselector.cxx
@@ -72,7 +72,7 @@ public:
 private:
 
    st_app::AppParGroup & m_pars;
-   float m_ra, m_dec, m_rad, m_tmin, m_tmax;
+   double m_ra, m_dec, m_rad, m_tmin, m_tmax;
 
    std::vector<std::string> m_inputFiles;
    std::string m_outputFile;
